Safari-611.1.10/LayoutTests/fast/dom/HTMLLinkElement/sizes-setter.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests that the HTMLLinkElement.sizes setter updates the associated sizes attribute");

var link = document.createElement("link");
shouldBe("link.sizes.__proto__", "DOMTokenList.prototype");
evalAndLog("link.sizes = '10x10  20x20'");
shouldBeEqualToString("String(link.sizes)", "10x10  20x20");
shouldBeEqualToString("link.getAttribute('sizes')", "10x10  20x20");
evalAndLog("link.sizes.add('30x30')");
shouldBeEqualToString("String(link.sizes)", "10x10 20x20 30x30");
shouldBeEqualToString("link.getAttribute('sizes')", "10x10 20x20 30x30");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
